% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LD.R
\docType{data}
\name{LD}
\alias{LD}
\title{LD Matrices for Top 10 Height-Associated Loci}
\format{
A list of 10 numeric matrices. Each matrix is symmetric and corresponds to one genomic locus.
}
\source{
UK Biobank; height GWAS summary statistics; PLINK clumping and LD estimation.
}
\usage{
data(LD)
}
\description{
This dataset is a list of estimated linkage disequilibrium (LD) matrices, one for each of the top 10 loci associated with height.
}
\details{
The loci were selected based on genome-wide significant associations (P < 5e-8) from a height GWAS. Clumping was performed using PLINK's clump-and-threshold (C+T) method with the following parameters:
\code{--clump-field P}, \code{--clump-kb 1000}, \code{--clump-p1 5e-8}, \code{--clump-p2 5e-8}, and \code{--clump-r2 0.9}, retaining independent index SNPs within each locus.

LD matrices were estimated using a reference panel of 9,680 unrelated individuals (4,475 males and 5,205 females) randomly selected from the UK Biobank.

Each element in the list corresponds to one locus and contains a symmetric numeric matrix of pairwise LD values (typically R-squared).
}
\examples{
data(LD)
length(LD)  # Number of loci
dim(LD[[1]])  # LD matrix for the first locus
}
\keyword{GWAS}
\keyword{LD}
\keyword{datasets}
\keyword{height}
