% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuSiE_Inf.R
\name{SuSiE_Inf}
\alias{SuSiE_Inf}
\title{Estimation of Causal Variants Using SuSiE-Inf}
\usage{
SuSiE_Inf(
  z,
  R,
  n,
  L = 5,
  pip.cred.thres = 0.95,
  max.iter = 50,
  max.eps = 0.001,
  susie.iter = 500,
  reml.iter = 10,
  score.test = T,
  pip.thres = 0.5,
  pv.thres = 0.05,
  eigen.thres = 1,
  varinf.upper.boundary = 0.25
)
}
\arguments{
\item{z}{Z scores of GWAS effect sizes.}

\item{R}{LD (Linkage Disequilibrium) matrix of variants.}

\item{n}{Sample size of GWAS data.}

\item{L}{The number of single effect used in SuSiE.}

\item{pip.cred.thres}{A threshold to define the credible set, default is 0.95.}

\item{max.iter}{Maximum number of iterations for estimating the infinitesimal effect, default is 15.}

\item{max.eps}{The maximum epsilon for convergence in the iterative process, default is 0.001.}

\item{susie.iter}{Number of iterations for the inner loop of the SuSiE, default is 500.}

\item{reml.iter}{Number of iterations for the inner loop of the REML, default is 10.}

\item{score.test}{Perform score test of variance component or not, default to F.}

\item{pip.thres}{In score test, a threshold to determine which variants are considered causal, default is 0.5. This will only be used in score test as a fixed effect adjustment.}

\item{pv.thres}{In score test, a threshold in score test of variance, default to 0.05. If the P-value of score test is larger than this threshold, the infinitesimal effect is removed.}

\item{eigen.thres}{The threshold of eigenvalues for modelling the infinitesimal effect. Default is 1.}

\item{varinf.upper.boundary}{The upper boundary for the prior variance of infinitesimal effects, multiplied by var(y) to adapt to different locus variances. Default is 0.25.}
}
\value{
A list containing the following elements:
  - \code{eta}: Linear predictor, which is the sum of the causal effect (\code{beta}) and the infinitesimal effect (\code{alpha}). It represents the total genetic effect of variants on the trait, combining the direct effects of causal variants and the background genetic effects.
  - \code{beta}: Causal effect. These are the effect estimates for variants identified with significant effect sizes through the SuSiE method, representing the direct impact of these variants on the trait.
  - \code{alpha}: Infinitesimal effect. This represents the aggregate effect of all other genetic variants in the background, excluding the identified causal variants. It can be considered as the unexplained genetic variability in the model.
  - \code{var.inf}: Variance of the infinitesimal effect. Estimated via REML, this reflects the degree of variation in the infinitesimal effect across the genetic variation.
  - \code{pv}: P-value from the score test. This value is used to assess whether the variance of the infinitesimal effect is significantly non-zero, serving as an important metric to test if the model adequately explains the genetic variability.
  - \code{pip}: Posterior Inclusion Probabilities for the estimates of causal effects. These are the probabilities calculated by the SuSiE method for each variant, assessing the likelihood that a variant is causal.
  - \code{fit.susie}: Output from the SuSiE method. This is the complete result object from the SuSiE method after performing fine-mapping, containing detailed information about the selected model and estimated parameters, allowing for further analysis and interpretation of the results.
}
\description{
This function utilizes the SuSiE (Sum of Single Effects) method for fine-mapping to identify causal variants with significant effect sizes. It employs REML (Restricted Maximum Likelihood) for the estimation of infinitesimal effects and for determining the optimal variance of these effects. Additionally, the function utilizes a score test to assess whether the variance of the infinitesimal effect is zero. The function has been updated to include parameters for controlling the precision of the estimation process and the option to estimate residual variance.
}
